task-#4 added native sqlite bindings and prepared integration tests
