+++ app/src/main/AndroidManifest.xml
@@ -1,13 +1,5 @@
-<manifest package="ru.arturvasilov.sqlite"
-          xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest package="ru.arturvasilov.sqlite">
 
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-
-    </application>
+    <application/>
 
 </manifest>
