+++ app/src/main/java/ru/arturvasilov/sqlite/core/BaseTable.java
@@ -1,8 +1,9 @@
 package ru.arturvasilov.sqlite.core;
 
-import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
 import android.support.annotation.NonNull;
+
+import org.sqlite.database.sqlite.SQLiteDatabase;
 
 /**
  * @author Artur Vasilov
