+++ app/src/main/java/ru/arturvasilov/sqlite/core/SQLiteContentProvider.java
@@ -4,12 +4,12 @@ import android.content.ContentProvider;
 import android.content.ContentUris;
 import android.content.ContentValues;
 import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
 import android.net.Uri;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.text.TextUtils;
+
+import org.sqlite.database.sqlite.SQLiteDatabase;
 
 /**
  * This class provides implementation for all operations in ContentProvider
@@ -24,7 +24,7 @@ public abstract class SQLiteContentProvider extends ContentProvider {
 
     private SQLiteSchema mSchema;
 
-    private SQLiteOpenHelper mSQLiteHelper;
+    private SQLiteHelper mSQLiteHelper;
 
     private static String sContentAuthority;
     private static Uri sBaseUri;
