+++ app/src/test/java/ru/arturvasilov/sqlite/utils/TableBuilderTest.java
@@ -1,117 +1,93 @@
 package ru.arturvasilov.sqlite.utils;
 
-import android.database.sqlite.SQLiteDatabase;
-
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import org.mockito.Mockito;
 
-import ru.arturvasilov.sqlite.testutils.TestTable;
+import ru.arturvasilov.sqlite.testutils.JUnitTestTable;
 
-import static org.mockito.Matchers.anyString;
+import static junit.framework.TestCase.assertEquals;
 
 @RunWith(JUnit4.class)
 public class TableBuilderTest {
 
-    private SQLiteDatabase mDb;
-
-    @Before
-    public void setUp() throws Exception {
-        mDb = Mockito.mock(SQLiteDatabase.class);
-        Mockito.doNothing().when(mDb).execSQL(anyString());
-    }
-
     @Test(expected = IllegalStateException.class)
     public void testEmptyColumnsList() throws Exception {
-        TableBuilder.create(TestTable.TABLE).execute(mDb);
+        TableBuilder.create(JUnitTestTable.TABLE).buildSQL();
     }
 
     @Test
     public void testSingleIntegerColumn() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(test INTEGER);";
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(test INTEGER);";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("test")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testSingleRealColumn() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(test REAL);";
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(test REAL);";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .realColumn("test")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testSingleStringColumn() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(test TEXT);";
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(test TEXT);";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .textColumn("test")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testPrimaryKeyWithIntColumn() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(test INTEGER, PRIMARY KEY (test));";
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(test INTEGER, PRIMARY KEY (test));";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("test")
                 .primaryKey("test")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testManyColumnsWithoutPrimaryKey() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(int1 INTEGER, int2 INTEGER, " +
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(int1 INTEGER, int2 INTEGER, " +
                 "real1 REAL, string1 TEXT, string2 TEXT);";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("int1")
                 .textColumn("string1")
                 .intColumn("int2")
                 .realColumn("real1")
                 .textColumn("string2")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testManyColumnsWithPrimaryKey() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(int1 INTEGER, int2 INTEGER," +
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(int1 INTEGER, int2 INTEGER," +
                 " real1 REAL, string1 TEXT, string2 TEXT, PRIMARY KEY (int1));";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("int1")
                 .textColumn("string1")
                 .intColumn("int2")
                 .realColumn("real1")
                 .textColumn("string2")
                 .primaryKey("int1")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testManyColumnsWithMultiplePrimaryKey() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(int1 INTEGER, int2 INTEGER, " +
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(int1 INTEGER, int2 INTEGER, " +
                 "real1 REAL, real2 REAL, string1 TEXT, string2 TEXT, PRIMARY KEY (int1, real1, string2));";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("int1")
                 .realColumn("real1")
                 .textColumn("string1")
@@ -119,36 +95,30 @@ public class TableBuilderTest {
                 .textColumn("string2")
                 .realColumn("real2")
                 .primaryKey("int1", "real1", "string2")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testSameColumnsIgnored() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(testInt INTEGER, testReal REAL, testText TEXT);";
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(testInt INTEGER, testReal REAL, testText TEXT);";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("testInt")
                 .intColumn("testInt")
                 .realColumn("testReal")
                 .realColumn("testReal")
                 .textColumn("testText")
                 .textColumn("testText")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 
     @Test
     public void testSamePrimaryKeysIgnored() throws Exception {
-        String sql = "CREATE TABLE IF NOT EXISTS TestTable(test INTEGER, PRIMARY KEY (test));";
+        String sql = "CREATE TABLE IF NOT EXISTS JUnitTestTable(test INTEGER, PRIMARY KEY (test));";
 
-        TableBuilder.create(TestTable.TABLE)
+        assertEquals(sql, TableBuilder.create(JUnitTestTable.TABLE)
                 .intColumn("test")
                 .primaryKey("test", "test")
-                .execute(mDb);
-
-        Mockito.verify(mDb).execSQL(sql);
+                .buildSQL());
     }
 }
