+++ build.gradle
@@ -14,6 +14,13 @@ buildscript {
     }
 }
 
+allprojects {
+    repositories {
+        jcenter()
+        flatDir { dirs 'libs' }
+    }
+}
+
 ext {
     PUBLISH_GROUP_ID = 'ru.arturvasilov'
     PUBLISH_ARTIFACT_ID = 'sqlite'
